Rework post discussions including removing IMessageTemplate
Still need to resolve issues noted by

- `//TODO: Pull 602- Fix this `
